More modes for build CLI
Support for dryrun, force build, and save tagged image

+++ run.sh
@@ -2,8 +2,28 @@
 
 set -o errexit
 
-WORK_DIR=$(dirname $0)
+# Check args
+while [[ $# -ge 1 ]]; do
+    case $1 in
+        -d|--dryrun)   # Dryrun (build but don't upload)
+            DRYRUN=1;;
+        -s|--save)  # Save the tagged image, don't delete
+            SAVE=1;;
+        -f|--force) # Force upload to overwrite the remote registry
+            FORCE=1;;
+        *)
+            echo "Usage: run.sh [OPTION]"
+            echo -e "\n Options:"
+            echo -e "\t-d, --dryrun\tDryrun (build the image but don't upload)"
+            echo -e "\t-s, --save\tSave the tagged image (don't delete the artifacts)"
+            echo -e "\t-f, --force\tForce the uplaod to overwrite the remote registry"
+            exit 1
+    esac
+    shift
+done
 
+# Set globals
+WORK_DIR=$(dirname $0)
 IMAGE_NAME="rdall96/minecraft-server"
 
 # Check docker version
@@ -33,7 +53,7 @@ docker manifest inspect "$IMAGE_NAME:$LATEST_VER" > /dev/null
 FOUND="$(echo $?)"
 set -o errexit
 
-if [[ $FOUND == 0 ]]; then
+if [[ $FOUND -eq 0 && $FORCE -ne 1 ]]; then
     # The image exists abort
     echo -e "\nFound image for latest Minecraft version in DockerHub '$IMAGE_NAME:$LATEST_VER', no need to rebuild"
     exit 0
@@ -50,12 +70,16 @@ else
     docker tag "$IMAGE_NAME:$MINECRAFT_VER" "$IMAGE_NAME:latest"
 
     # Push to DockerHub
-    echo -e "\nPushing to DockerHub..."
-    docker push "$IMAGE_NAME:$MINECRAFT_VER"
-    docker push "$IMAGE_NAME:latest"
+    if [[ $DRYRUN -ne 1 ]]; then
+        echo -e "\nPushing to DockerHub..."
+        docker push "$IMAGE_NAME:$MINECRAFT_VER"
+        docker push "$IMAGE_NAME:latest"
+    fi
 fi
 
 # Cleanup
 deactivate
 rm -rf "$PYTHON_VENV"
-docker rmi "$IMAGE_NAME:$MINECRAFT_VER" "$IMAGE_NAME:latest"
+if [[ $SAVE -ne 1 ]]; then
+    docker rmi "$IMAGE_NAME:$MINECRAFT_VER" "$IMAGE_NAME:latest"
+fi
